# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/tasks/{id}": {
            "patch": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Updates a task",
                "description": "# Updates a task/subtask for authenticated user\n### Id\nYou must specify an id of a task you want to update in the URL.\n\n### Json body\nYou can specify such fields as: parent_id, priority, title, description.<br>\nAll fields are optional.\n\n### Parent id\nYou can specify id of a task you wish to inherit in a field 'parent_id' to attach this task to another one.<br>\nYou cannot specify parent_id of a task, that doesn't belong to you,<br>\nneither you can specify a descendant task id as a parent id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Specify an id of a task you wish to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "parent_id": {
                                        "type": "integer",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "priority": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1,
                                        "example": 4,
                                        "nullable": true
                                    },
                                    "title": {
                                        "type": "string",
                                        "maximum": 255,
                                        "example": "Clean the house",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "maximum": 3000,
                                        "example": "Clean everything in that house.",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 22
                                                    },
                                                    "userId": {
                                                        "type": "integer",
                                                        "example": null
                                                    },
                                                    "parentId": {
                                                        "type": "integer",
                                                        "example": null
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "todo"
                                                    },
                                                    "priority": {
                                                        "type": "integer",
                                                        "example": 4
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Clean the house"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Clean everything in that house."
                                                    },
                                                    "completedAt": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "createdAt": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "29-10-2022 20:44:30"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "If user is not logged in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden operations",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Not your task": {
                                        "summary": "Not your task",
                                        "value": {
                                            "message": "You don't own this task"
                                        }
                                    },
                                    "Parent is not your task": {
                                        "summary": "Parent is not your task",
                                        "value": {
                                            "message": "You cannot add subtask to the task of another user"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Undesired behaviour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object",
                                            "example": "You cannot attach task to its own subtask"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/tasks/{id}/delete": {
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Deletes user's task",
                "description": "# Deletes a task/subtask of authenticated user\n### Id\nYou must specify an id of a task you want to delete in the URL.\n\n### Constraints\nYou cannot delete a completed task.<br>\nYou cannot delete a task with completed subtasks.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Specify an id of a task you wish to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object",
                                            "example": "Task with id 22 successfully deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "If user is not logged in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden operations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object",
                                            "example": "You don't own this task"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Undesired behaviour",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "Delete a completed task": {
                                        "summary": "Delete a completed task",
                                        "value": {
                                            "message": "You cannot delete a completed task"
                                        }
                                    },
                                    "Delete a task with completed subtasks": {
                                        "summary": "Delete a task with completed subtasks",
                                        "value": {
                                            "message": "You cannot delete a task with completed subtasks"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}